examples/rust-matter-light: implement own Handler for OnOffCluster; move UdpSocketWrapper to network.rs
