+++ examples/rust-matter-light/src/network.rs
@@ -1,15 +1,105 @@
 use core::ops::DerefMut as _;
-use embedded_hal_async::delay::DelayNs;
 
-use rs_matter::{self, transport::network::{UdpReceive, UdpSend}};
+use rs_matter;
 use rs_matter::error::Error as MatterError;
-use riot_wrappers::socket_embedded_nal_async_udp::UnconnectedUdpSocket;
 use riot_wrappers::gnrc::Netif;
-use embedded_nal_async::{Ipv4Addr, Ipv6Addr, UdpStack as _, UnconnectedUdp};
+use embedded_nal_async::{Ipv4Addr, Ipv6Addr, SocketAddr, UdpStack as _, UnconnectedUdp};
+
+use embassy_futures::select::{Either, select};
+use riot_wrappers::mutex::Mutex;
+use riot_wrappers::ztimer;
+use riot_wrappers::socket_embedded_nal_async_udp::UnconnectedUdpSocket;
+use rs_matter::error::{Error, ErrorCode};
+use rs_matter::transport::network::{UdpReceive, UdpSend};
+use embassy_sync::{
+    signal::Signal,
+    blocking_mutex::raw::NoopRawMutex
+};
+use embedded_hal_async::delay::DelayNs;
+use log::{debug, error, warn};
+use riot_wrappers::error::NumericError;
+
+pub type Notification = Signal<NoopRawMutex, ()>;
+
+pub struct UdpSocketWrapper {
+    local_addr: SocketAddr,
+    socket: Mutex<UnconnectedUdpSocket>,
+    release_socket_notification: Notification,
+    socket_released_notification: Notification,
+}
+
+impl UdpSocketWrapper {
+    pub fn new(local_addr: SocketAddr, socket: UnconnectedUdpSocket) -> Self {
+        Self {
+            local_addr,
+            socket: Mutex::new(socket),
+            release_socket_notification: Notification::new(),
+            socket_released_notification: Notification::new(),
+        }
+    }
+}
+
+impl UdpSend for &UdpSocketWrapper {
+    async fn send_to(&mut self, data: &[u8], addr: SocketAddr) -> Result<(), Error> {
+        debug!("(UDP) sending {} bytes to {:?}", data.len(), &addr);
+        if addr.is_ipv4() {
+            // IPv4 not supported on RIOT OS
+            warn!("(UDP) Is IPv4 -> ignore send!");
+            return Ok(());
+        }
+        // tell receiver to release mutex within max. 10ms
+        self.release_socket_notification.signal(());
+        ztimer::Delay.delay_ms(5).await;
+        let mut sock = self.socket.try_lock().expect("receiver should have ensured that this mutex is free");
+        sock.send(self.local_addr, addr, data)
+            .await
+            .map_err(|_| Error::new(ErrorCode::StdIoError))?;
+        // tell receiver sending is finished
+        drop(sock);
+        self.socket_released_notification.signal(());
+        Ok(())
+    }
+}
+
+impl UdpReceive for &UdpSocketWrapper {
+    async fn recv_from(&mut self, buffer: &mut [u8]) -> Result<(usize, SocketAddr), Error> {
+        loop {
+            let mut sock = self.socket.try_lock().expect("sender should have ensured that this mutex is free");
+            match select(
+                self.release_socket_notification.wait(),
+                sock.receive_into(buffer)
+            ).await {
+                Either::First(_) => {
+                    // give up mutex for sender
+                    drop(sock);
+                    // ... and wait until available again
+                    self.socket_released_notification.wait().await;
+                    continue;
+                }
+                Either::Second(res) => {
+                    match res {
+                        Ok((bytes_recvd, local_addr, remote_addr)) => {
+                            debug!("(UDP) received {} bytes from {:?} to {:?}", bytes_recvd, &remote_addr, &local_addr);
+                            if remote_addr.is_ipv4() {
+                                warn!("(UDP) Is IPv4 -> ignoring receive data!");
+                                return Ok((bytes_recvd, remote_addr));
+                            }
+                        }
+                        Err(_) => { error!("Error during UDP receive!"); }
+                    }
+                    // return receive result
+                    let (bytes_recvd, remote_addr) = res.map(|(bytes_recvd, _, remote_addr)|
+                        (bytes_recvd, remote_addr)
+                    ).map_err(|_| Error::new(ErrorCode::StdIoError))?;
+                    return Ok((bytes_recvd, remote_addr));
+                }
+            }
+        }
+    }
+}
 
 pub mod utils {
     use log::{debug, error, info};
-    use crate::socket::UdpSocketWrapper;
     use super::*;
 
     pub fn get_ipv6_address(ifc: &Netif) -> Option<Ipv6Addr> {
